requirements.txt
numpy
scipy
six
mpi4py
tensorflow-gpu>=1.4
cached_property
python-dateutil
tqdm
matplotlib
